refactor: move `consensus_threshold` function in utils
